refactor: change nat traversal module to torrente
