+++ src/torrente/NatTraversalHandler.ts
@@ -1,6 +1,6 @@
 import portControl from 'nat-puncher';
-import { PAYFLUXO_LISTENING_PORT, PAYFLUXO_EXTERNAL_PORT } from '../../config';
-import { NotificationTypesEnum } from '../../torrente/notification/models/NotificationModel';
+import { PAYFLUXO_LISTENING_PORT, PAYFLUXO_EXTERNAL_PORT } from '../config';
+import { NotificationTypesEnum } from './notification/models/NotificationModel';
 
 const handleNatTraversalProbing = async () =>{
     portControl.addMapping(
